+++ ProcessFiles.h
@@ -1,7 +1,7 @@
 /*
  * The MIT License
  *
- * Copyright 2017 Totoros.
+ * Copyright 2017 Theomeli.
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -24,11 +24,10 @@
 
 /* 
  * File:   ProcessFiles.h
- * Author: Totoros
+ * Author: Theomeli
  *
  * Created on June 5, 2017, 1:43 PM
  */
-
 
 #ifndef PROCESSFILES_H
 #define PROCESSFILES_H
@@ -47,87 +46,106 @@ public:
     virtual ~ProcessFiles();
     
     /**
-     * 
-     * @return 
+     * getter for private member documentsText
+     * @return the ifstream documentsText
      */
 	ifstream& getDocumentsText() { return documentsText; }
     
     /**
-     * 
-     * @return 
+     * getter for private member queriesText
+     * @return the ifstream queriesText
      */
 	ifstream& getQueriesText() { return queriesText; }
     
     /**
-     * getter for documentsTokens
+     * getter for private member documentsTokens
      * @return the list of documentsTokens
      */
 	list<string>* getDocumentsTokens() const { return documentsTokens; }
     
     /**
-     * getter for queriesTokens
+     * getter for private member queriesTokens
      * @return the list of queriesTokens
      */
 	list<string>* getQueriesTokens() const { return queriesTokens; }
     
     /**
-     * 
-     * @return 
+     * getter for private member nDocuments
+     * @return the number of documents
      */
-	int getNDocuments() const { return nDocuments; }
+	size_t getNDocuments() const { return nDocuments; }
     
     /**
-     * 
-     * @return 
+     * getter for private member nQueries
+     * @return the number of queries
      */
-	int getNQueries() const { return nQueries; }
+	size_t getNQueries() const { return nQueries; }
     
     /**
-     * 
-     * @return 
+     * getter for private member nResponses
+     * @return the number of responses
      */
-	map<int, int> getNResponses() const { return nResponses; }
+	map<size_t, size_t> getNResponses() const { return nResponses; }
     
     /**
-     * 
-     * @param stream 
+     * it reads the documents file. Firstly stores the number of documents
+	 * to variable nDocuments and then documents' tokens to variable 
+	 * documentsTokens according to documents' ids
+     * @param stream is the stream to be read
      */
     void readDocumentsFile(ifstream& stream);
     
     /**
-     * 
-     * @param stream
+	 * it reads the queries file. Firstly stores the number of queries to 
+	 * variable nQueries. Consequently stream gives two numbers to be read
+	 * which stored to map nResponses and finally queries' tokens which are
+	 * stores to variable queriesTokens according to queries' ids
+     * @param stream is the stream to be read
      */
     void readQueriesFile(ifstream& stream);
 private:
-    ifstream documentsText;
-    ifstream queriesText;
-    int nDocuments;
-    int nQueries;
-    list<string>* documentsTokens;
-    list<string>* queriesTokens;
-    //key: the query id, value: the number of responses will be returned for this query
-    map <int, int> nResponses;
+	//input file stream for the documents 
+    ifstream									documentsText;
+	//input file stream for the queries
+    ifstream									queriesText;
+	//number of documents
+	//TODO: nDocuments and nQueries to be a 
+	//variable map<bool, size_t> nDocs so that 
+	//the bool part to answer the query isQuery
+	size_t										nDocuments;
+	//number of queries
+	size_t										nQueries;
+	//a list for each document containing its 
+	//terms
+	//TODO: replace pointer with a vector
+    list<string>*								documentsTokens;
+	//a list for each query containing its terms
+	//TODO: replace pointer with a vector
+    list<string>*								queriesTokens;
+    //key: the query id, value: the number of 
+	//responses that will be returned for this 
+	//query
+	map <size_t, size_t>						nResponses;
     
     /**
-     * 
-     * @param s
-     * @return 
+     * it transforms a string to its lower case
+     * @param s the string to be transformed
+     * @return the trasnformed string
      */
     string makeLower(const string& s);
     
     /**
-     * 
-     * @param s
-     * @param punct
-     * @return 
+     * it removes the punctuation from a string 
+     * @param s the string to be checked for punctuation
+     * @param punct the string of punctuation's characters
+     * @return the string without punctuation
      */
     string removePunct(const string& s, const string& punct);
     
     /**
-     * 
-     * @param s
-     * @return 
+     * it uses the functions makeLower and removePunct
+     * @param s the string to be transformed
+     * @return a string without punctuation and in a lower case
      */
     string lowerRemovedPunct(const string& s);
 };
